refactor: split keymap.rs into keymap.rs + mode_hints.rs
keymap.rs was 1101 lines. Extracted mode_keymap() function and all
per-mode hint/fkey data into mode_hints.rs (619 lines), leaving
keymap.rs at 468 lines with core types, global_keybindings(),
fit_* functions, and tests.
